without points!

+++ light.h
@@ -5,9 +5,9 @@
 #include "geom.h"
 class PointLightSource{
 public:
-    PointLightSource(ThreeDVector pos_, double power_) : pos(pos_), power(power_) {};
+    PointLightSource(ThreeDVector pos_, long double power_) : pos(pos_), power(power_) {};
     ThreeDVector pos;
-    double power; //в канделах
+    long double power; //в канделах
 };
 
 #endif // LIGHT_H
